+++ lib/db/parsed-statement.js
@@ -5,7 +5,7 @@
 
 'use strict'
 
-const _recordMetrics = require('../../lib/metrics/recorders/record-metrics')
+const _recordMetrics = require('../../lib/metrics/recorders/database')
 
 function ParsedStatement(type, operation, collection, raw) {
   this.type = type
