+++ lib/metrics/recorders/database-operation.js
@@ -5,7 +5,7 @@
 
 'use strict'
 
-const metrics = require('../../metrics/names')
+const metrics = require('../names')
 
 /**
  * Records all the metrics required for database operations.
