+++ lib/metrics/recorders/database.js
@@ -5,7 +5,7 @@
 
 'use strict'
 
-const { DB, ALL } = require('../../metrics/names')
+const { DB, ALL } = require('../names')
 const { DESTINATIONS } = require('../../config/attribute-filter')
 
 /**
