+++ lib/shim/datastore-shim.js
@@ -15,7 +15,7 @@ const Shim = require('./shim')
 const urltils = require('../util/urltils')
 const util = require('util')
 const specs = require('./specs')
-const recordOperationMetrics = require('../../lib/metrics/recorders/record-operation-metrics')
+const recordOperationMetrics = require('../../lib/metrics/recorders/database-operation')
 const { DatastoreParameters } = specs.params
 
 /**
