+++ lib/shim/webframework-shim/middleware.js
@@ -13,7 +13,7 @@ const {
 } = require('./common')
 const { assignCLMSymbol } = require('../../util/code-level-metrics')
 const { RecorderSpec } = require('../specs')
-const makeMiddlewareRecorder = require('../../metrics/recorders/middleware-metric-recorder')
+const makeMiddlewareRecorder = require('../../metrics/recorders/middleware')
 
 const MIDDLEWARE_TYPE_DETAILS = {
   APPLICATION: { name: 'Mounted App: ', path: true, record: false },
